# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "orqion",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/providers/{provider_id}/models": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Create Model",
        "operationId": "create_model_api_providers__provider_id__models_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/models/{model_id}": {
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Update Model",
        "operationId": "update_model_api_providers_models__model_id__patch",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Delete Model",
        "description": "Удаление модели провайдера — метаданные (T-443, коммит 2).\n\nЕсли модель закреплена за корпусом (corpus.pinned_model_id) — 409 без\nкаскада, иначе гарантия ADR-12 терялась бы тихо. Проверка до delete:\nSELECT count(*) FROM corpus WHERE pinned_model_id = :id.\n\nОпциональная очистка с диска (delete_from_disk=true) — только по явному\nподтверждению из UI, не поведение по умолчанию; проводится через\nнативные эндпоинты контракта model_download.py с тем же kind-гейтом\n(DOWNLOADABLE_KINDS). Ошибка диска НЕ блокирует удаление метаданных,\nно возвращается в ответе и показывается пользователю.\n\nСсылки из истории (message.model_id, usage_event.model_id) обнуляются —\nпрецедент usage_event «при удалении диалога связи становятся NULL»:\nзаписи о сообщениях и расходе сохраняются, модель отвязывается.\nСтроки бессрочного агрегата usage_daily переводятся на сентинел NIL_ID\n(BUG-019): обнуление невозможно (model_id в составе PK), удаление\nпротиворечило бы §5.3 — агрегат переживает первичные данные.",
        "operationId": "delete_model_api_providers_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "delete_from_disk",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Delete From Disk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelCreate": {
        "properties": {
          "alias": {
            "type": "string",
            "title": "Alias"
          },
          "upstream_name": {
            "type": "string",
            "title": "Upstream Name"
          },
          "locality": {
            "type": "string",
            "title": "Locality",
            "default": "local"
          },
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "supports_reasoning": {
            "type": "boolean",
            "title": "Supports Reasoning",
            "default": false
          },
          "reasoning_toggleable": {
            "type": "boolean",
            "title": "Reasoning Toggleable",
            "default": false
          },
          "cost_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost In"
          },
          "cost_out": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Out"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "alias",
          "upstream_name"
        ],
        "title": "ModelCreate"
      },
      "ModelDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "disk_deleted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Deleted"
          },
          "disk_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disk Error"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "ModelDeleteResponse",
        "description": "Результат удаления модели провайдера (T-443, коммит 2).\n\ndisk_deleted / disk_error заполняются только при запрошенной очистке\nс диска; ошибка диска не блокирует удаление метаданных."
      },
      "ModelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "alias": {
            "type": "string",
            "title": "Alias"
          },
          "upstream_name": {
            "type": "string",
            "title": "Upstream Name"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "provider_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Kind"
          },
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "supports_reasoning": {
            "type": "boolean",
            "title": "Supports Reasoning"
          },
          "reasoning_toggleable": {
            "type": "boolean",
            "title": "Reasoning Toggleable"
          },
          "cost_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost In"
          },
          "cost_out": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Out"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "alias",
          "upstream_name",
          "locality",
          "max_input_tokens",
          "max_output_tokens",
          "supports_reasoning",
          "reasoning_toggleable",
          "cost_in",
          "cost_out",
          "enabled"
        ],
        "title": "ModelResponse"
      },
      "ModelUpdate": {
        "properties": {
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias"
          },
          "upstream_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Name"
          },
          "locality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality"
          },
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "supports_reasoning": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Reasoning"
          },
          "reasoning_toggleable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Toggleable"
          },
          "cost_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost In"
          },
          "cost_out": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Out"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ModelUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}